requirements.txt
python-dotenv
google-genai
google

torch
torchaudio
transformers
accelerate
openai-whisper
soundfile
